# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PAD Lazuardy API",
        "description": "API documentation for PAD Lazuardy backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/student/review": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "List tutors to review",
                "operationId": "8e8bc564da00a56c186367d0d8b9de78",
                "responses": {
                    "200": {
                        "description": "Review list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Students"
                ],
                "summary": "Update review",
                "operationId": "5ac0da24eafa38a19f01c6c908079574",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tutor_id",
                                    "quality",
                                    "delivery",
                                    "attitude",
                                    "benefit",
                                    "rate"
                                ],
                                "properties": {
                                    "tutor_id": {
                                        "type": "integer"
                                    },
                                    "quality": {
                                        "type": "string"
                                    },
                                    "delivery": {
                                        "type": "string"
                                    },
                                    "attitude": {
                                        "type": "string"
                                    },
                                    "benefit": {
                                        "type": "string"
                                    },
                                    "rate": {
                                        "type": "integer"
                                    },
                                    "review": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/student/review/{tutor_id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get review detail",
                "operationId": "f42cb8de09c2dc7759d9b4fec53796d5",
                "parameters": [
                    {
                        "name": "tutor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-packages": {
            "get": {
                "tags": [
                    "StudyPackages"
                ],
                "summary": "Get student's purchased packages",
                "operationId": "f4f07338d4b66caef5a77c6f630a9908",
                "responses": {
                    "200": {
                        "description": "Packages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/study-packages": {
            "get": {
                "tags": [
                    "StudyPackages"
                ],
                "summary": "Get study packages",
                "operationId": "3c965fa48521616d31663b511a2aa3fb",
                "responses": {
                    "200": {
                        "description": "Packages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/study-packages/{id}": {
            "get": {
                "tags": [
                    "StudyPackages"
                ],
                "summary": "Get study package detail",
                "operationId": "8ff1720f3e37a51af60f87e859e1f7f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/find-tutor": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Find tutors",
                "operationId": "2faf0df016d5dd92f4e56425c9b5ac84",
                "parameters": [
                    {
                        "name": "radius",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_rating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "weight_rating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "weight_distance",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tutors list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/find-tutor/{id}": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Find tutor detail",
                "operationId": "b28c72922c4e4660b6e5e031afd3796c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tutor detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StandardSuccess": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token from login/register responses",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}